Implement SMTP config check tests and refactor for testability
- Refactor contact API route logic into app/api/contact/logic.ts to enable unit testing.
- Add app/api/contact/route.test.ts using Node.js built-in test runner.
- Add npm test script to package.json.
- Ensure 500 error is returned when SMTP credentials are missing.
